requirements.txt
Django==2.2.7
django-rest==0.0.1
dnspython==1.16.0
pytz==2019.3
sqlparse==0.3.0
